Add serialization, make name integral to class
It's important for the name to have a concrete link to the task as it'll
need to know what name to give to the serialized data, without needing
to rely on a mapping.

Add serialize methods to these classes, but for now leave them
unimplemented; I'll fill those in next.

Add canSerialize to allow the serializer to specify whether or not it
can work on a particular condition. This is required because:
  * Type erasure makes it difficult to identify the correct serializer,
    as we won't know what our T is
  * It will allow more complex logic for serialization. Though for now
    it's a little boilerplatey as each implementation of the method just
    checks the type against its concrete T, to avoid type erasure, it's
    possible future implementations might analyse the condition further
    and be capable of serialization multiple types of condition.