Add tests for StringConditions serializers
While I'm here, fix a bug with StringConditions which meant that the
appliesTo(String) method could not be reached if appliesTo was called on
a Condition representing the SCondition. I thought I'd tested that
worked regardless, but I guess not well enough.

Add a test to cover the above

Add a couple of TODOs to add features which could be useful to these
conditions, too.