Add REST interface to the switches
Can be served as a spray API if desired
A couple of endpoints aren't yet complete

To be able to properly finish this API to check if a switch is active,
more serialization work will need to be done as the provided context to
the "is_active" endpoint will have to be deserialized according to
extensible rules such that it can be compared to switches properly

The same needs to be done for the "equals" rule, really. For now we'll
just support basic types like String, Int, Double, etc., same as the
"equals" comparator.

Abstract this functionality into a ContextValue serializer
Add a DefaultFormats object to make things easier to use when
customisation isn't required