Add a predictive "proportional" switch
Use SHA-1 to ensure consistency, and compare the hash of data provided
to proportion * max sha1 value (2^160) to see if it's in the correct
proportion.

This provides a somewhat ridiculous resolution at the expense of
efficiency, so it'll probably be worth stripping that down to only the
first few bytes in future.