Remove createSwitch
Doesn't seem necessary; so far it's a noop in both backends, and
probably would be in every other. configureSwitch lets you specify a
name and config, so it's immaterial if it exists or not.

Might replace this later with an 'update_only' flag or something which
will prevent conflicts, but that'll be a different mechanism.