Tests for Proportion and fix the logic
Needed to either compare the full hash, to the correct max hash
value, or only take the first n bytes of the generated hash. I've
adjusted this to only look at the first 10 octets - that's more than
enough to establish a proportion with a fair degree of resolution.

Also commit the is_hodor_strict.json fixture which was left uncommitted
from a previous commit, whoops.