listActive should be more efficient
Made a brainfart when I wrote this, clearly: I took a list of tuples
containing switch names and configs, and then used isActive with the
switch name to check if the switches were active... which, of course,
refetches the configs I've already gone to the trouble of fetching.

Modify the default listActive implementation so that it doesn't do that.
Move TODO on the redis backend so that the efficiency note is where it
is actually needed. I've already mostly written the batched MGET code,
so I'll try that out soon, too.