Add StringConditions.OneOf
This is a less verbose way of providing a list of strings to match:
otherwise it'd have to be an AND containing lots of EQUALS conditions,
and that could get very large to achieve something which is a fairly
likely case to want to handle.

Add serialisation for this class
Add tests for both the class and the serialisation

While I'm here, set tests to execute in parallel and print full stack
traces of errors.

Also skip the redis tests, again, because once again
whisklabs/docker-it-scala has broken my test suite.