Issue #20: Warn when leaving unsaved changes
Maintain a list of switches on Admin for this purpose, ensuring
SwitchBuilders are saved there when created.

Add a method to SwitchBuilder to allow figuring out if the switch is
saved or not. I wasn't updating 'condition' on them when they were
updated, which I should have been anyway, really. If the current state
of the form either fails to validate or renders as JSON different to the
current saved condition, it's unsaved.

Bind to onbeforeunload (with a no conflict option to disable that if
desired) to make sure we look for unsaved switches and warn users
they're going to discard changes if they leave.