Admin: refactor cancel -> self.reset, call on save
Triggering this on save will ensure the widget is updated to the state
it'll be persisted as. This includes replacing any raw JSON fields with
properly-build widgets, instead.

Also change updateSwitch to take the condition as js objects, not a JSON
string: relic from the textarea business; it was just converting it back
into an object anyway.

Fix a case where the save button doesn't become re-enabled

Make sure widgets which construct subconditions always use
createCondition rather than instantiating directly