Re-enable redis tests
They work as long as you don't have dangling docker images. A fix went
into RC1 of the whisklabs lib to fix that problem, so I'll upgrade to
RC1 soon. Involves a couple of fiddles with extra deps and adding a
missing method to make it work, so not as simple as just upgrading,
sadly.